Change data type

+++ gateway/payloads/events/threadmembersupdate.go
@@ -8,7 +8,7 @@ import (
 type ThreadMembersUpdate struct {
 	ThreadId         uint64                  `json:"id,string"`
 	GuildId          uint64                  `json:"guild_id,string"`
-	MemberCount      uint8                   `json:"member_count"`
+	MemberCount      uint64                  `json:"member_count"`
 	AddedMembers     []channel.ThreadMember  `json:"added_members"`
 	RemovedMemberIds utils.Uint64StringSlice `json:"removed_member_ids"`
 }
